disable remote fetching in Paperclip
There is a CVE[1] in Paperclip v5.1.0, the version Supermarket depends on
as of writing this.

[1] https://nvd.nist.gov/vuln/detail/CVE-2017-0889

***** Supermarket is not vulnerable to this CVE. *****

The `tarball` parameter given for an upload is validated by
CookbookUpload::Parameters and CookbookUpload::Archive before a hand-off
to ActiveRecord+Paperclip for processing as an attachment.

This update to the Paperclip initializer is intended to prevent future
accidental support for remote fetching if the logic of CookbookUpload
changes at some point and allows a URL string through to the persistence
layer.

Signed-off-by: Robb Kidd <rkidd@chef.io>